ussc/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-rest-swagger==2.2.0
django-mptt
psycopg2